refs swagger-api/swagger-core#2312 - resolve string with format, cleanup
